requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
deprecation==2.1.0
Flask==3.1.0
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Levenshtein==0.26.1
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.0
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
RapidFuzz==3.10.1
realtime==2.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.0
